requirements.txt
# Core
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# Database
pymongo==4.5.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# NLP & ML
torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2
spacy==3.7.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

# RAG & Vector Database
langchain==0.0.350
faiss-cpu==1.7.4
chromadb==0.4.18

# PDF Processing
pypdf==3.17.0
pdfplumber==0.10.3
python-docx==1.1.0

# Data Processing
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# APIs
requests==2.31.0
groq==0.3.0
openai==1.3.0

# Development
black==23.11.0
flake8==6.1.0
pytest==7.4.3